Backend/requirements.txt
fastapi[all]
sqlalchemy
pydantic
python-dotenv
asyncpg



